Added leave and entry points/adjusted move function
Made the car follow the full path and added multiple points to the botLeft intersection.
