+++ Assets/Materials/New Material 1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 647155dbadc57cc45a7f114d1767c5f5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
